requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0
lxml==4.9.3
tqdm==4.66.1
openpyxl==3.1.2